BUG/MEDIUM: lib: account for empty id/name for alerts/detections
It was recently pointed out that there was an issue with Templated Rules
(specifically virtual patches) that was causing terraform updates made
after a manual UI update to not be properly reflected in the UI.

Reproduction is as follows:

1. Enable a Templated Rule using Terraform.
2. Navigate to the Templated Rule and "Disable" it from the UI
3. Re-run Terraform to enable the rule
4. Navigate to the list of Templated Rules within the UI and search for
the Templated Rule in question. It will show as "Disabled".

It was also found that a similar issue was happening when just making
the updates directly in Terraform without involving the UI. This could
be resolved by re-running the plan several times. In both cases
described it was found that the API was actually returning the following
error as part of it's JSON response:

```
	"detectionErrors": ["add {\"id\":\"\",\"name\":\"CVE-2018-9206\",\"enabled\":true,\"fields\":[],\"created\":\"\",\"createdBy\":\"\"} Validation failed"]
```

This error was actually part of a successful API response, which means
changes were partially being applied.

After a bit of investigation it was found that when Terraform was doing
an update it was sending only detectionAdds/detectionDeletes and
alertAdds/alertDeletes and detectionUpdates/alertUpdates was null. This
seems to be due to a comparison that was happening between the "old"
and "new" values, which was comparing "ID" and "Name" values, which
had the potential to be empty within the "new" value.

This commit fixes both issues described above by checking if "new" has
an empty ID or Name when the "old" value is not empty. We then update
the "new" ID/Name values to be the same as the "old" one for each of
the alert/detection add/delete/update functions.